删除textContent，大地图添加关卡

+++ TypeScript/Game/Component/UndergroundComponent.ts
@@ -48,7 +48,9 @@ export class UndergroundComponent extends Component implements IUndergroundCompo
         if (isChildOfClass(other.Entity.Actor, TsPlayer)) {
             this.RestartPlayerPos();
         } else {
-            gameContext.EntityManager.RemoveEntity(other);
+            if (other.Entity.Actor.ActorHasTag('DestroyUnderground')) {
+                gameContext.EntityManager.RemoveEntity(other);
+            }
         }
     }
